PUSH

+++ engine/src/utils.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: alde-fre <alde-fre@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/11/25 09:38:43 by alde-fre          #+#    #+#             */
+/*   Updated: 2022/11/25 09:38:44 by alde-fre         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "engine.h"
 
 int	ft_min(int a, int b)
@@ -23,8 +35,7 @@ void	ft_swap(void **a, void **b)
 	*b = c;
 }
 
-
 float	ft_rand(float min, float max)
 {
-	return ((max - min) * ((float)rand() / RAND_MAX) + min);
+	return ((max - min) * ((float)rand() / (float)RAND_MAX) + min);
 }
